+++ 2020_RTSS_cpc_model_based_algorithm/src/main.rs
@@ -4,9 +4,9 @@ mod prioritization_cpc_model;
 use clap::Parser;
 use lib::fixed_priority_scheduler::FixedPriorityScheduler;
 use lib::homogeneous::HomogeneousProcessor;
+use lib::log::dump_dag_scheduler_result_to_yaml;
 use lib::processor::ProcessorBase;
 use lib::scheduler::DAGSchedulerBase;
-use lib::util::dump_dag_scheduler_result_to_yaml;
 use lib::{dag_creator::*, graph_extension::GraphExtension};
 use log::warn;
