+++ 2021_RTCSA_dynfed/src/main.rs
@@ -5,11 +5,10 @@ use dynfed::DynamicFederatedScheduler;
 use lib::dag_creator::*;
 use lib::fixed_priority_scheduler::FixedPriorityScheduler;
 use lib::homogeneous::HomogeneousProcessor;
+use lib::log::dump_dag_set_scheduler_result_to_yaml;
 use lib::processor::ProcessorBase;
 use lib::scheduler::DAGSetSchedulerBase;
-use lib::util::{
-    adjust_to_implicit_deadline, dump_dag_set_scheduler_result_to_yaml, get_hyper_period,
-};
+use lib::util::{adjust_to_implicit_deadline, get_hyper_period};
 
 #[derive(Parser)]
 #[clap(
@@ -47,10 +46,5 @@ fn main() {
 
     let file_path = dynfed_scheduler.dump_log(&arg.output_dir_path, "FixedPriority");
 
-    dump_dag_set_scheduler_result_to_yaml(
-        &file_path,
-        schedule_length,
-        get_hyper_period(&dag_set),
-        schedule_length < get_hyper_period(&dag_set),
-    );
+    dump_dag_set_scheduler_result_to_yaml(&file_path, schedule_length < get_hyper_period(&dag_set));
 }
