+++ lib/src/util.rs
@@ -85,48 +85,6 @@ pub fn create_yaml(folder_path: &str, file_name: &str) -> String {
     file_path
 }
 
-#[derive(Serialize, Deserialize)]
-struct DAGSchedulerResultInfo {
-    schedule_length: i32,
-    period_factor: f32,
-    result: bool,
-}
-
-pub fn dump_dag_scheduler_result_to_yaml(
-    file_path: &str,
-    schedule_length: i32,
-    period_factor: f32,
-    result: bool,
-) {
-    let result_info = DAGSchedulerResultInfo {
-        schedule_length,
-        period_factor,
-        result,
-    };
-    dump_struct(file_path, &result_info);
-}
-
-#[derive(Serialize, Deserialize)]
-struct DAGSetSchedulerResultInfo {
-    schedule_length: i32,
-    hyper_period: i32,
-    result: bool,
-}
-
-pub fn dump_dag_set_scheduler_result_to_yaml(
-    file_path: &str,
-    schedule_length: i32,
-    hyper_period: i32,
-    result: bool,
-) {
-    let result_info = DAGSetSchedulerResultInfo {
-        schedule_length,
-        hyper_period,
-        result,
-    };
-    dump_struct(file_path, &result_info);
-}
-
 #[cfg(test)]
 mod tests {
     use super::*;
